Make delayms an uint16

+++ ledsquare/LED_Animation.h
@@ -13,7 +13,7 @@ public:
     void loop(void);
     virtual void animation(void);
     uint16_t step;
-    uint8_t delayms;
+    uint16_t delayms;
     uint16_t _VIEW_HEIGHT, _VIEW_WIDTH;
     int16_t _sVIEW_HEIGHT, _sVIEW_WIDTH;
     LED_Strip *_led;
